fix: 여행지 검색창 제거
