+++ src/pages/SearchPlacePage/index.jsx
@@ -1,15 +1,15 @@
 import { useParams } from "react-router-dom";
 import PageContainer from "../../components/PageContainer";
 import PlaceMap from "../../components/PlaceMap";
+import Title from "../../components/Title";
 
 function SearchPlacePage() {
   const { keyword } = useParams();
 
   return (
     <PageContainer>
-      <p className="text-2xl text-black font-bold my-9">
-        가고 싶은 여행지를 찾아요
-      </p>
+      <Title title="가고 싶은 여행지를 찾아요" />
+      <div className="my-7" />
       <PlaceMap
         search={true}
         keyword={keyword}
